requirements.txt
Django==2.1.5
Pillow==5.4.1
python-memcached=1.12.0